pyproject.toml
[tool.poetry]
name = "ldraney-ebay-sdk"
version = "0.1.0"
description = "Python SDK wrapping all eBay REST APIs"
authors = ["ldraney"]
readme = "README.md"
license = "MIT"
packages = [{include = "ebay_sdk", from = "src"}]
keywords = ["ebay", "api", "sdk", "ecommerce"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = "^3.10"
ldraney-ebay-oauth = ">=0.1.0"
httpx = ">=0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0"
pytest-asyncio = ">=0.23"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (hit real eBay sandbox API)",
]
testpaths = ["tests"]
